requirements.txt
asgiref==3.8.1
boto3==1.35.10
botocore==1.35.10
Django==5.1
django-allauth==64.2.0
django-storages==1.14.4
djangorestframework==3.15.2
grpcio==1.65.5
grpcio-tools==1.65.5
jmespath==1.0.1
pillow==10.4.0
protobuf==5.27.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
